fix bugs : 1. null pointer in vk 2. pause if download cover art 3. fragment exception

+++ src/com/perm/DoomPlay/DownloadingService.java
@@ -52,7 +52,7 @@ public class DownloadingService extends Service implements Download.DoomObserver
         switch (holder.download.getStatus())
         {
             case DOWNLOADING:
-                notification = holder.downloadBuilder.createStarting((int)holder.download.getProgress());
+                notification = holder.downloadBuilder.createStarting(holder.download.getProgress());
                 startUpdatingThread();
                 break;
             case CANCELLED:
